+++ problems/p10/p10.mojo
@@ -17,7 +17,7 @@ comptime layout = Layout.row_major(SIZE, SIZE)
 def shared_memory_race(
     output: LayoutTensor[dtype, layout, MutAnyOrigin],
     a: LayoutTensor[dtype, layout, ImmutAnyOrigin],
-    size: UInt,
+    size: Int,
 ):
     var row = thread_idx.y
     var col = thread_idx.x
@@ -45,7 +45,7 @@ def shared_memory_race(
 def add_10_2d(
     output: LayoutTensor[dtype, layout, MutAnyOrigin],
     a: LayoutTensor[dtype, layout, ImmutAnyOrigin],
-    size: UInt,
+    size: Int,
 ):
     var row = thread_idx.y
     var col = thread_idx.x
@@ -94,7 +94,7 @@ def main() raises:
             ctx.enqueue_function[add_10_2d, add_10_2d](
                 out_tensor,
                 a_tensor,
-                UInt(SIZE),
+                SIZE,
                 grid_dim=BLOCKS_PER_GRID,
                 block_dim=THREADS_PER_BLOCK,
             )
@@ -123,7 +123,7 @@ def main() raises:
             ctx.enqueue_function[shared_memory_race, shared_memory_race](
                 out_tensor,
                 a_tensor,
-                UInt(SIZE),
+                SIZE,
                 grid_dim=BLOCKS_PER_GRID,
                 block_dim=THREADS_PER_BLOCK,
             )
